lbrce file is the forntend files of our project and the backend file has the source of backend codes.
This project is mainly in a static form cause it has dynamic mode for only the registration for logins and for op forms only and for rest of the project we only used the staticc method. If want you can download it and start the execution from the signup page no need of changes to make.
Make sure if want change any name in the code you must check in other codes no make changes as the files are connected to each other.
The execution flow of this project is you need to signup with valid details where these details will saved into the mongoDB, then you can login after login you may the diseases page here you need to select your particular disease, when you selected the disease it will navigate to the hospitals page (we have connected same hospitals for several diseases), after checking the hospital details you can select you particular hospital it will redirect to the doctors page(we have connected same doctors for several hospitals), there you can select your doctor when you click on doctor, you will get a opform and you need to fill the form with valid details and need to submit the form after submitting, you will greeted with thank you message.

This is the execution flow.